fix warning on OpenBSD

+++ test/time/test_boot.c
@@ -8,6 +8,6 @@ int main(void)
 	time_t boot = neb_time_boot();
 	if (!boot)
 		return -1;
-	fprintf(stdout, "boottime: %ld seconds since epoch time\n", boot);
+	fprintf(stdout, "boottime: %lld seconds since epoch time\n", (long long)boot);
 	return 0;
 }
